Added native parser to Dining Hall Menu Feature
The dining hall menu feature no longer relies on an outside JSON file in
order to read in menu data. Using JSOUP, the menu information is now
parsed directly from the app. Right now only the meal name is displayed,
but the plan is to add a field to select meal times (breakfast, lunch,
dinner) next.